refactor(stores): split store.svelte.ts into domain modules (#295)
- Create src/lib/stores/ with domain files:
  - ui-stores.svelte.ts (Modal, Query, Toast, MainControls, FloatingControlPanel)
  - map-stores.svelte.ts (Map, MapView, MapTools, Terrain, Building3D)
  - editor-stores.svelte.ts (EditorChrome, MapEdit, MapProposal, AdditionProposal, EventPlacement)
  - sync-stores.svelte.ts (AppBootstrap, SyncToast, Offline)
  - data-stores.svelte.ts (Term, RoomClasses)
- Coupled stores (Location, Jeepney, AdminAuth, Proposals, ScheduleRoute) stay in index.ts
- store.svelte.ts is now a thin barrel: export * from './stores/index.js'
- All @lib/store.svelte imports work unchanged
